app/FakeNewsDetectorAPI/requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-filter==24.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
scikit-learn==1.5.0
requests==2.32.3
gunicorn==21.2.0
whitenoise==6.6.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
google-generativeai==0.3.0